requirements.txt
aiobotocore==1.1.1
aiohttp==3.6.2
aioitertools==0.7.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.2
black==19.10b0
bleach==3.2.1
boto3==1.15.6
botocore==1.17.44
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.1
fsspec==0.8.3
great-expectations==0.12.2
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
lxml==4.5.2
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
newspaper3k==0.2.8
nltk==3.5
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==9.0.1
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.9.27
requests==2.23.0
requests-file==1.5.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3fs==0.5.1
s3transfer==0.3.3
scipy==1.5.2
Send2Trash==1.5.0
setuptools==39.0.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
